backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
python-multipart==0.0.9
pdfplumber==0.11.0
pandas==2.2.1
openpyxl==3.1.2
openai==1.16.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
xlrd==2.0.1
# ── Auth & Security ──────────────────────────────────────────────
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
aiosmtplib==3.0.1
itsdangerous==2.2.0
